"Refactor scripts and update Python version"
This commit encompasses both modification of the python scripts and updates on the Python version. In the '2a_amiga.py' file, unused imports and unnecessary code blocks were removed improving readability and efficiency. In the '1b_hello.py' file, additional functionalities were added that include a new spinCameraTask method, a sequence to coordinate intervals for the panda's walk and update on panda's position. Also in multiple notebooks, the Python version was updated from "3.11.3" to "3.9.7" likely to match the environment's Python version. The installation command for Panda3D in notebooks was also updated to include an upgrade flag.