Improved tutorial slide structure and updated to recent build platforms
Adjusted the content structure in the Panda3D tutorial 2 by changing slide_type to 'subslide' for improved clarity. Also updated the build platforms to 'manylinux2014_x86_64', to reflect necessity in working with the latest versions. The Python version in the notebooks has been updated to "3.11.5" for compatibility with latest features and libraries. In addition to this, numerous enhancements have been made to the 1c_tutorial_1.py script for better organization and efficiency. Furthermore, enabled Black formatter plugin in the Misc.xml file for better code quality assurance.